refactor: extract AI tools to dedicated @repo/chat-ai-tools package
This commit creates a new `@repo/chat-ai-tools` package to house the
webSearch and createDocument tool definitions, separating them from
the core chat types.

## Changes Made

### New Package: @repo/chat-ai-tools
- Created dedicated package for AI tool definitions
- Moved `LightfastAppChatToolSet` and related types from chat-core
- Includes tool definitions for `webSearch` and `createDocument`
- Proper TypeScript configuration with ESNext module support

### Updated @repo/chat-core
- Removed tool definitions from core package
- Added dependency on @repo/chat-ai-tools
- Re-exports tool types for backward compatibility
- Keeps existing imports working without changes

### Benefits
- Better separation of concerns
- chat-core no longer needs to infer tool types
- AI tools are now in a dedicated, focused package
- Maintains backward compatibility with existing imports
- Cleaner dependency graph

## Result
All builds pass and existing functionality is preserved while improving
the package structure and separation of concerns.
